Added the 3 required cogs & the ability to load/unload sub-cogs
Added the cogs Reloader, Navi, and Journal.

I have also added the ability to load/unload sub-cogs so parts so a full cog can be removed if unwanted.
This also changed the way the listcogs command worked and will now show the extension first then the cogs in that extension.

Still todo is move the SQL for these cogs