Prevent attempts at nesting charsets in regex filter
Not sure how I didn't catch this when testing a couple years ago.
Basically, nesting character sets is impossible in Python's regex, which
is what my filter code tried to do because it wasn't checking for IN
tokens. This adds that check and performs confusable insertion in place
for those sets. Also cleans up some of the code in that area.